+++ packages/api/src/API/Http/Actions/User/EditUserAction.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from 'inversify';
 import EditUserAdapter from '../../Adapter/User/EditUserAdapter';
 import EditUserHandlerInterface from '../../../../Infraestructure/Interfaces/User/EditUserHandlerInterface';
 import EditUserCommand from '../../../../Application/Commands/User/EditUserCommand';
-import TYPES from '../../../../Infraestructure/types';
+import TYPES from '../../../../Infraestructure/DI/types';
 
 @injectable()
 class EditUserAction {
