+++ packages/api/src/API/Http/Adapter/Product/ShowAllProductAdapter.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
-import { InvalidData } from '../../Errors/InvalidData';
+import { InvalidData } from '../../Errors/BadRequest';
 import IdSchema from '../../Validator/Schemas/IdSchema';
 import ProductFindCommand from '../../../../Application/Commands/Product/ProductFindCommand';
 import { injectable } from 'inversify';
