+++ packages/api/src/API/Http/Adapter/UserRole/CreateUserRoleAdapter.ts
@@ -2,22 +2,22 @@ import { Request, Response } from 'express';
 import { inject, injectable } from 'inversify';
 import Validator from '../../Validator/Validator';
 import { BadRequest } from '../../Errors/BadRequest';
-import DeleteUserRoleCommand from '../../../../Application/Commands/UserRole/DeleteUserRoleCommand';
-import { IdSchema } from '../../Validator/Schemas/Common';
+import CreateUserRoleCommand from '../../../../Application/Commands/UserRole/CreateUserRoleCommand';
+import { CreateUserRoleSchema } from '../../Validator/Schemas/UserRoleSchema';
 
 @injectable()
-class DeleteUserRoleAdapter {
+class CreateUserRoleAdapter {
   private validator: Validator;
   constructor(@inject(Validator) validator: Validator) {
     this.validator = validator;
   }
-  public async from(req: Request): Promise<DeleteUserRoleCommand> {
-    const error = this.validator.validate(req.params, IdSchema);
+  public async from(req: Request): Promise<CreateUserRoleCommand> {
+    const error = this.validator.validate(req.body, CreateUserRoleSchema);
     if (error) {
       throw new BadRequest(JSON.stringify(this.validator.validationResult(error)));
     }
-    return new DeleteUserRoleCommand(req.body);
+    return new CreateUserRoleCommand(req.body);
   }
 }
 
-export default DeleteUserRoleAdapter;
+export default CreateUserRoleAdapter;
