+++ packages/api/src/API/Http/Adapter/UserRole/EditUserRoleAdapter.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from 'inversify';
 import Validator from '../../Validator/Validator';
 import { BadRequest } from '../../Errors/BadRequest';
 import EditUserRoleCommand from '../../../../Application/Commands/UserRole/EditUserRoleCommand';
-import EditUserRoleSchema from '../../Validator/Schemas/UserRoleSchema';
+import { EditUserRoleSchema } from '../../Validator/Schemas/UserRoleSchema';
 import { IdSchema } from '../../Validator/Schemas/Common';
 
 @injectable()
@@ -21,7 +21,7 @@ class EditUserRoleAdapter {
     if (errorId) {
       throw new BadRequest(JSON.stringify(this.validator.validationResult(errorId)));
     }
-    return new EditUserRoleCommand(req.body);
+    return new EditUserRoleCommand(Number(req.params.id), req.body.name);
   }
 }
