+++ packages/api/src/API/Http/Adapter/UserRole/FindUserRoleAdapter.ts
@@ -2,22 +2,22 @@ import { Request, Response } from 'express';
 import { inject, injectable } from 'inversify';
 import Validator from '../../Validator/Validator';
 import { BadRequest } from '../../Errors/BadRequest';
-import CreateUserRoleCommand from '../../../../Application/Commands/UserRole/CreateUserRoleCommand';
-import CreateUserRoleSchema from '../../Validator/Schemas/UserRoleSchema';
+import FindUserRoleCommand from '../../../../Application/Commands/UserRole/FindUserRoleCommand';
+import { FindUserRoleSchema } from '../../Validator/Schemas/UserRoleSchema';
 
 @injectable()
-class CreateUserRoleAdapter {
+class FindUserRoleAdapter {
   private validator: Validator;
   constructor(@inject(Validator) validator: Validator) {
     this.validator = validator;
   }
-  public async from(req: Request): Promise<CreateUserRoleCommand> {
-    const error = this.validator.validate(req.body, CreateUserRoleSchema);
+  public async from(req: Request): Promise<FindUserRoleCommand> {
+    const error = this.validator.validate(req.params, FindUserRoleSchema);
     if (error) {
       throw new BadRequest(JSON.stringify(this.validator.validationResult(error)));
     }
-    return new CreateUserRoleCommand(req.body);
+    return new FindUserRoleCommand(req.body.name);
   }
 }
 
-export default CreateUserRoleAdapter;
+export default FindUserRoleAdapter;
