+++ packages/api/src/API/Http/Errors/BadRequest.ts
@@ -1,6 +1,6 @@
 import { InfraestructureError } from '../../../Infraestructure/Errors/InfraestructureError';
 
-export class InvalidData extends InfraestructureError {
+export class BadRequest extends InfraestructureError {
   constructor(message: string) {
     super(message, 400);
   }
