+++ packages/api/src/Application/Handlers/Auth/ChangePasswordHandler.ts
@@ -1,8 +1,8 @@
 import { injectable, inject } from 'inversify';
 import ChangePasswordCommand from '../../Commands/Auth/ChangePasswordCommand';
 import IUserRepository from '../../../Domain/Interfaces/IUserRepository';
-import TYPES from '../../../Infraestructure/types';
-import { EntityNotFound } from '../../../API/Http/Errors/EntityNotFound';
+import TYPES from '../../../Infraestructure/DI/types';
+import { EntityNotFound } from '../../../Infraestructure/Errors/EntityNotFound';
 import { UnAuthorizedError } from '../../../API/Http/Errors/UnAuthorizedException';
 import User from '../../../Domain/Entities/User';
