+++ packages/api/src/Domain/Enums/UserRoles.ts
@@ -1,0 +1,5 @@
+export enum UserRoles {
+  Zeep = 'Zeep',
+  Admin = 'Admin',
+  Owner = 'Owner',
+}
