+++ packages/api/src/routes/index.routes.ts
@@ -1,4 +1,5 @@
 import userRoutes from './user.routes';
+import userRoleRoutes from './userrole.routes';
 import productRoutes from './product.routes';
 import authRoutes from './auth.routes';
 import { Router } from 'express';
@@ -6,6 +7,8 @@ import { Router } from 'express';
 const router = Router();
 
 router.use(productRoutes);
+
+router.use(userRoleRoutes);
 
 router.use(userRoutes);
